pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ashlr-ao"
dynamic = ["version"]
description = "Local-first agent orchestration platform — one developer, many AI coding agents, single command center."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
authors = [{ name = "Mason Wyatt" }]
keywords = ["ai", "agents", "orchestration", "claude", "codex", "tmux"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: MIT License",
    "Topic :: Software Development :: Build Tools",
    "Typing :: Typed",
    "Framework :: aiohttp",
]
dependencies = [
    "aiohttp>=3.9.0",
    "psutil>=5.9.0",
    "pyyaml>=6.0",
    "aiosqlite>=0.19.0",
    "bcrypt>=4.0.0",
    "PyJWT[crypto]>=2.8.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-asyncio>=0.23.0", "pytest-aiohttp>=1.0.0", "ruff>=0.4.0", "pytest-cov>=4.0"]

[project.scripts]
ashlr = "ashlr_ao.server:main"

[project.urls]
Homepage = "https://github.com/Ashlar-inc/ashlar-ao"
Repository = "https://github.com/Ashlar-inc/ashlar-ao"

[tool.setuptools.dynamic]
version = { attr = "ashlr_ao.__version__" }

[tool.setuptools.packages.find]
include = ["ashlr_ao*"]

[tool.setuptools.package-data]
ashlr_ao = ["dashboard.html", "logo.png"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
